Bilinear filtering + Perspective corrected texturemapping implemented - but not enabled in the rendering pipeline. Need to refactor a bit of things before they can be part of the fun.. :)

+++ Source/Balder/Rendering/Silverlight/RenderVertex.cs
@@ -75,9 +75,9 @@ namespace Balder.Rendering.Silverlight
 
 		public void MakeScreenCoordinates()
 		{
-			TranslatedScreenCoordinates.X = TranslatedVector.X;
-			TranslatedScreenCoordinates.Y = TranslatedVector.Y;
-			TranslatedScreenCoordinates.Z = TranslatedVector.Z;
+			TranslatedScreenCoordinates.X = (int)TranslatedVector.X;
+			TranslatedScreenCoordinates.Y = (int)TranslatedVector.Y;
+			TranslatedScreenCoordinates.Z = (int)TranslatedVector.Z;
 		}
 
 		public Vector TransformedVector;
